Finish decoupling UnitAura collection from frame updates
Unify the logic for Classic and Retail in how we collect unit aura information. Importantly, we now have unit aura collection decoupled from the frame visual updates. This should be a performance win for Classic, albeit much less dramatic than on Retail with the new UNIT_AURA payloads added in 10.0.