Code cleanup to database, core processing loop, and addon loading
This big code cleanup will lead to some minor performance improvements and some major quality of life changes in the code. This version does mandate a database migration to the new spec, meaning you will not be able to revert to an older version of ERF once updating to this version.