Rewrite our method of hiding the stock aura frames
This new way of hiding the stock buff, debuff, and dispel debuff frames should be more reliable and efficient, as we're now secure hooking the OnShow callback functions directly rather than relying on hooking CompactUnitFrame_UpdateAuras(). This should hopefully suppress the random boss auras that show up in raids, and should also be more efficient as a bonus effect.