Make sure that we create the listener regardless of frame visibility
We were hitting an issue where the aura listener could sometimes not get created, leading to a broken experience. This was because we added some extra checks prior to creating the aura listener. However, we want the listener no matter what, so instead we restrict our checks to solely checking whether the unit matches player, party, or raid strings.